+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ unreleased
 * Revert change to the toolbar sites menu to use ``http`` protocol.
 * Fix edit plugin popup width (remove 850px width constraint).
 * Fix except block using list instead of tuple. (#7334)
+* Added cache ttl extension point.
 
 3.10.0 (2022-03-26)
 ===================